-se realizó registro estudiante y curso completo

+++ tarea01/src/controlador/Controlador_JFrame_Curso.java
@@ -8,6 +8,7 @@ package controlador;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import vista.JFrame_ConsultaCurso;
 import vista.JFrame_Curso;
 import modelo.MetodosCurso;
 
@@ -18,10 +19,12 @@ import modelo.MetodosCurso;
 public class Controlador_JFrame_Curso implements ActionListener{
 
     //Referencia 
+    JFrame_ConsultaCurso consultaCurso;
     JFrame_Curso curso;
     MetodosCurso metodos;
     
     public Controlador_JFrame_Curso(JFrame_Curso curso) {
+        consultaCurso = new JFrame_ConsultaCurso();
         this.curso=curso;
         metodos = new MetodosCurso(curso);
     }
@@ -29,26 +32,81 @@ public class Controlador_JFrame_Curso implements ActionListener{
     public void actionPerformed(ActionEvent e){
         
         if(e.getActionCommand().equals("Registro")){
-            metodos.registrarCurso(curso.obtenerInformacionCurso());
-            curso.limpiarCampos();
-        }
+            
+            if(!curso.obtenerSiglaCurso().equals("")){
+                
+                if(!metodos.siExisteCurso(curso.obtenerSiglaCurso())){
+                    
+                    if(!curso.obtenerSiglaCurso().equals("") && !curso.obtenerNombreCurso().equals("") && !curso.obtenerCreditoCurso().equals("")){
+                        
+                        metodos.registrarCurso(curso.obtenerInformacionCurso());
+                        curso.limpiarCampos();
+                        
+                    }else{
+                        
+                        metodos.mensajeFaltaUnCampo();
+                        
+                    }
+                }else{
+                    
+                    metodos.mensajeAtencion();
+                    curso.limpiarCampos();
+                }
+            }else{
+                
+                metodos.mensajeAtencion();
+                
+            }
+        }//Fin Registro
         
         if(e.getActionCommand().equals("Modificar")){
-            metodos.modificarCurso();
-            curso.limpiarCampos();
-        }
+            
+            if(metodos.siExisteCurso(curso.obtenerSiglaCurso())){
+                
+                metodos.modificarCurso();
+                curso.limpiarCampos();    
+                
+            }else{
+                
+                metodos.mensajeNoExisteEstudiante();
+                curso.limpiarCampos();
+                
+            }
+            
+        }//Fin Modificar
         
         if(e.getActionCommand().equals("Eliminar")){
-            metodos.eliminarCurso(curso.obtenerSiglaCurso());
-            curso.limpiarCampos();
-        }
+            if(metodos.siExisteCurso(curso.obtenerSiglaCurso())){
+                
+                metodos.eliminarCurso(curso.obtenerSiglaCurso());
+                curso.limpiarCampos();  
+                
+            }else{
+                
+                metodos.mensajeNoExisteEstudiante();
+                curso.limpiarCampos();
+                
+            }
+            
+        }//Fin Eliminar
         
-        if(e.getActionCommand().equals("Buscar")){       
-            metodos.consultarCurso(curso.obtenerSiglaCurso());
-        }
+        if(e.getActionCommand().equals("Buscar")){
+            
+            if(!curso.obtenerSiglaCurso().equals("")){
+                
+                metodos.consultarCurso(curso.obtenerSiglaCurso());
+                
+            }else{
+                
+                metodos.mensajeFaltaUnCampo();
+                
+            }
+            
+        }//Fin Buscar
         
         if(e.getActionCommand().equals("Consultar")){
-            metodos.obtenerListaCursos();
+            consultaCurso.setVisible(true);            
+            consultaCurso.cargarInformacionCursos(metodos.obtenerListaCursos());
         }
         
     }//Fin Controlador_JFrame_Curso
